// my-openapi-zod-app/src/api/Schema/ZProductUpcAllocationModel.ts
//
// This file is generated. Do not edit manually.
//
// @applicationName: GreenOnion.API
// @applicationDescription: GreenOnion.API - 1.0.0.0
// @applicationVersion: 1.0.0.0
// @generatedDate: 2025/07/26
//

import { z } from "zod";

export const ZProductUpcAllocationModel = z.object({
    productId:z.number().int().optional(),
    upc:z.string().optional(),
    isActive:z.boolean().optional(),
    createdBy:z.number().int().optional(),
});